+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Update marge on `ImageCropper` component.
+- Fix: Update marge on `FormComposer` component.
 
 ## [16.2.0] - 2017-11-22