+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Change `quinary` to `senary` modifier of `Title` on `ProjectCard` and
   `ProjectSimilarCard`.
 - Feature: Add `senary` modifier of `Title` component.
+- Fix: Update marge on `FormComposer` component.
 
 ## [16.2.0] - 2017-11-22