+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Update `focusId` on `Label` component.
 - Fix: Add `line-height` on `searchable` of `Select` component.
 - Fix: Update marge on `FormComposer` component.