+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Change `quinary` to `senary` modifier of `Title` on `ProjectCard` and
   `ProjectSimilarCard`.
 - Feature: Add `senary` modifier of `Title` component.
+- Feature: Update `focusId` on `Label` component.
+- Fix: Add `line-height` on `searchable` of `Select` component.
 - Fix: Update marge on `FormComposer` component.
 
 ## [16.2.0] - 2017-11-22