+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `k-u-anti-aliasing` helper mixin.
 - Fix: Push empty object when selection is removed in `SelectWithState`.
 - Feature: Update `focusId` on `Label` component.
 - Fix: Add `line-height` on `searchable` of `Select` component.