+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `instagram` modifier in `buttonIcon` component,
-  `InstagramButtonIcon` and `InstagramIcon` component.
+  `InstagramButtonIcon` and `InstagramIcon` components.
 - Feature: Add `k-buttonPadding` helper to manage button paddings.
 - Feature: Update all `Button` with new design styleguide V2.
 - Fix: Update marge on `FormComposer` component.