+++ CHANGELOG.md
@@ -6,6 +6,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fix: Update margin on `ImageCropper` component.
 - Fix: Update margin on `FormComposer` component.
+- Fix: Push empty object when selection is removed in `SelectWithState`.
+- Feature: Update `focusId` on `Label` component.
+- Fix: Add `line-height` on `searchable` of `Select` component.
+- Fix: Update marge on `FormComposer` component.
 
 ## [16.2.0] - 2017-11-22