+++ CHANGELOG.md
@@ -5,7 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add `k-buttonPadding` helper to manage button paddings.
-- Feature: Update all `Button` with new design styleguide V2.
+- Feature: Update all `Button` components with new design styleguide V2.
 - Fix: Update marge on `FormComposer` component.
 
 ## [16.2.0] - 2017-11-22