+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Update margin on `ImageCropper` component.
+- Fix: Update margin on `FormComposer` component.
 - Feature: Add `k-u-anti-aliasing` helper mixin.
 - Fix: Push empty object when selection is removed in `SelectWithState`.
 - Feature: Update `focusId` on `Label` component.