+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Change `quinary` to `senary` modifier of `Title` on `ProjectCard` and
+  `ProjectSimilarCard`.
+- Feature: Add `senary` modifier of `Title` component.
 - Fix: Update margin on `ImageCropper` component.
 - Fix: Update margin on `FormComposer` component.
 - Feature: Add `k-u-anti-aliasing` helper mixin.