+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Fix: Fix `null` value in `SelectWithState` component.
+
 ## [16.3.0] - 2017-11-27
 
 Features: